refactor: 소셜로그인은 `oauth`로 일반로그인은 `auth` 디렉토리로 이동
